bitcoind: Fix blockHash ZMQ test
The test mines blocks and waits until it sees the block hashes of the
mined blocks over ZMQ. Unfortunately, ZMQ publishers are stateless and
can't re-transmit lost messages [1]. Consequently, the test was wrong.
The fixed test subscribes to the publisher, starts mining blocks, and
waits until it observes a new block hash.

[1] https://zguide.zeromq.org/docs/chapter5/#Pros-and-Cons-of-Pub-Sub